+++ examples/ansibleplaybook-embed-python/internal/ansibleplaybook-embed-python/generate/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"github.com/kluctl/go-embed-python/pip"
+)
+
+func main() {
+	err := pip.CreateEmbeddedPipPackagesForKnownPlatforms("requirements.txt", "./data/")
+	if err != nil {
+		panic(err)
+	}
+}